Rename RekeyDriver and move extension functions
