+++ library/driver/api/jvm/driver.api
@@ -19,14 +19,14 @@ public final class io/toxicity/sqlite/mc/driver/MCConfigQueries$Companion {
 	public final fun from (Lio/toxicity/sqlite/mc/driver/SQLiteMCDriver;)Lio/toxicity/sqlite/mc/driver/MCConfigQueries;
 }
 
-public abstract class io/toxicity/sqlite/mc/driver/RekeyDriver : app/cash/sqldelight/driver/jdbc/JdbcDriver, app/cash/sqldelight/db/SqlDriver {
-	protected static final field Companion Lio/toxicity/sqlite/mc/driver/RekeyDriver$Companion;
-	public synthetic fun <init> (Lio/toxicity/sqlite/mc/driver/RekeyDriver$Companion$Args;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+public abstract class io/toxicity/sqlite/mc/driver/PlatformDriver : app/cash/sqldelight/driver/jdbc/JdbcDriver, app/cash/sqldelight/db/SqlDriver {
+	protected static final field Companion Lio/toxicity/sqlite/mc/driver/PlatformDriver$Companion;
+	public synthetic fun <init> (Lio/toxicity/sqlite/mc/driver/PlatformDriver$Companion$Args;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun addListener ([Ljava/lang/String;Lapp/cash/sqldelight/Query$Listener;)V
 	public final fun beginTransaction (Ljava/sql/Connection;)V
 	public final fun close ()V
 	public final fun closeConnection (Ljava/sql/Connection;)V
-	public static final synthetic fun create$driver (Lio/toxicity/sqlite/mc/driver/config/FactoryConfig;Ljava/util/Map;Z)Lio/toxicity/sqlite/mc/driver/RekeyDriver$Companion$Args;
+	public static final synthetic fun create$driver (Lio/toxicity/sqlite/mc/driver/config/FactoryConfig;Ljava/util/Map;Z)Lio/toxicity/sqlite/mc/driver/PlatformDriver$Companion$Args;
 	public final fun currentTransaction ()Lapp/cash/sqldelight/Transacter$Transaction;
 	public final fun endTransaction (Ljava/sql/Connection;)V
 	public final fun execute (Ljava/lang/Integer;Ljava/lang/String;ILkotlin/jvm/functions/Function1;)Lapp/cash/sqldelight/db/QueryResult;
@@ -41,13 +41,13 @@ public abstract class io/toxicity/sqlite/mc/driver/RekeyDriver : app/cash/sqldel
 	public final fun setTransaction (Lapp/cash/sqldelight/driver/jdbc/ConnectionManager$Transaction;)V
 }
 
-protected final class io/toxicity/sqlite/mc/driver/RekeyDriver$Companion {
+protected final class io/toxicity/sqlite/mc/driver/PlatformDriver$Companion {
 }
 
-public final class io/toxicity/sqlite/mc/driver/SQLiteMCDriver : io/toxicity/sqlite/mc/driver/RekeyDriver {
+public final class io/toxicity/sqlite/mc/driver/SQLiteMCDriver : io/toxicity/sqlite/mc/driver/PlatformDriver {
 	public final field config Lio/toxicity/sqlite/mc/driver/config/FactoryConfig;
 	public final field isInMemory Z
-	public synthetic fun <init> (Lio/toxicity/sqlite/mc/driver/config/FactoryConfig;ZLio/toxicity/sqlite/mc/driver/RekeyDriver$Companion$Args;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (Lio/toxicity/sqlite/mc/driver/config/FactoryConfig;ZLio/toxicity/sqlite/mc/driver/PlatformDriver$Companion$Args;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 }
 
 public final class io/toxicity/sqlite/mc/driver/SQLiteMCDriver$Factory {
