+++ library/driver/src/commonMain/kotlin/io/toxicity/sqlite/mc/driver/SQLiteMCDriver.kt
@@ -45,7 +45,7 @@ public class SQLiteMCDriver private constructor(
     @JvmField
     public val isInMemory: Boolean,
     args: Companion.Args
-): RekeyDriver(args) {
+): PlatformDriver(args) {
 
     /**
      * Primary entry point to create new [SQLiteMCDriver] instances.
